Move DCIM SVG code to `nautobot.dcim.svg` submodules
