+++ nautobot/dcim/svg/rack_elevation.py
@@ -5,8 +5,8 @@ import svgwrite
 
 from nautobot.core.utils.config import get_settings_or_config
 
-from .choices import DeviceFaceChoices
-from .constants import RACK_ELEVATION_BORDER_WIDTH
+from nautobot.dcim.choices import DeviceFaceChoices
+from nautobot.dcim.constants import RACK_ELEVATION_BORDER_WIDTH
 
 
 class RackElevationSVG:
